Reduce Code Reuse, Store Money in Integers
Now storing monetary values as integers to prevent any kind of rounding issues.

Moved certain functionality from Livewire components related to Agreement Items into a trait for reuse across components.

Moved Modal Blade into a single file with conditional rendering of components based on if it is currently editing an item or is creating a new item, reducing reuse enormously.

+++ database/migrations/2024_09_30_101110_create_agreement_items_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('name');
             $table->string('description');
             $table->integer('quantity');
-            $table->decimal('cost_price', 10, 2);
-            $table->decimal('retail_price', 10, 2);
+            $table->integer('cost_price');
+            $table->integer('retail_price');
             $table->timestamps();
             $table->softDeletes();
         });
